Add Zone model and serialization support
Established a standard way of doing json serialization to and from a class
structure for Zone.

This is not meant to be comprehensive, but to illustrate the pattern.

The file `serdes.py` has functions `to_json_string` and `from_json_string` that
handle all of the serialization.

Going `to_json_string` is handled primarily by the `to_dict` function that
converts any nested object structure to a dictionary.  Along the way, it will
convert field names in `snake_case` (with underscores) to field names in
`camelCase`.  This is important as it allows us to stay PEP compliant with our
field names, while maintaining interoperability with the VinylDNS API which uses
`camelCase`.

`from_json_string` is trickier to handle as there are no type hints in the JSON
structure coming back from the API.  Therefore, we need to pass in an
`object_ctor` function that knows how to take a `dict` and generate an object
instance.  Therefore, the actual `object_ctor` has to be known at the call site
(which is reasonable).

This leaves us with the model in `zone`.  The pattern here is to provide
**static method** `from_dict` on each class that takes a dictionary and yields a
new class instance.  Without type hinting in JSON, this is the only way we could
do this.  I looked into use `namedtuple` but they had their own issues that made
it difficult to do to and from `camelCase` so this
structure, while boilerplate and annoying, is small trade-off for deterministic
serialization.

The `test_vinyldns` has a simple test to show that the nested object structure
of a `Zone` does in fact serialize naturally to and from json using the
technique introduced in this PR.

Looking for feedback on this method.  Extending it to include the rest of the
VinylDNS API is a trivial / tedious task.

serialization.